feat: Implement value api: transactionByID
