+++ plugins/webapi/value/unspentoutputs/handler.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/iotaledger/goshimmer/dapps/valuetransfers"
 	"github.com/iotaledger/goshimmer/dapps/valuetransfers/packages/address"
+	"github.com/iotaledger/goshimmer/plugins/webapi/value/utils"
 	"github.com/labstack/echo"
 	"github.com/labstack/gommon/log"
 )
@@ -34,7 +35,7 @@ func Handler(c echo.Context) error {
 			if output.ConsumerCount() == 0 {
 				outputids = append(outputids, OutputID{
 					ID: id.String(),
-					InclusionState: InclusionState{
+					InclusionState: utils.InclusionState{
 						Confirmed: true,
 						Conflict:  false,
 						Liked:     true,
@@ -72,13 +73,6 @@ type UnspentOutput struct {
 
 // OutputID holds the output id and its inclusion state
 type OutputID struct {
-	ID             string `json:"id"`
-	InclusionState `json:"inclusion_state"`
-}
-
-// InclusionState represents the different states of an OutputID
-type InclusionState struct {
-	Confirmed bool `json:"confirmed, omitempty"`
-	Conflict  bool `json:"conflict, omitempty"`
-	Liked     bool `json:"liked, omitempty"`
+	ID             string               `json:"id"`
+	InclusionState utils.InclusionState `json:"inclusion_state"`
 }
